Move markup transition completion to the language server
`MarkupTransitionCompletionItemProvider` is only used in the language server, so this change moves it out of Workspaces
